fix: Miri compatibility and add Miri CI
- Change Block::base and Arena::cur_base from usize to *mut u8 for strict provenance
- Add zero-sized allocation checks in InlineVec to fix Miri UB
- Track block alignment and bypass fast path for high-alignment requests (>64)
- Add Miri CI job running on nightly with strict provenance
- Fix duplicate test code in tests/inline_vec.rs

All 162 tests pass. Benchmarks show no significant regression.

+++ src/util/transaction.rs
@@ -178,7 +178,7 @@ impl<'arena> Transaction<'arena> {
     /// total.
     #[inline(always)]
     pub fn bytes_used(&self) -> usize {
-        let current_used = self.arena.cur_ptr as usize - self.arena.cur_base;
+        let current_used = self.arena.cur_ptr as usize - self.arena.cur_base as usize;
         (self.arena.bytes_allocated + current_used).saturating_sub(self.bytes_at_open)
     }
